requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Django==3.0.5
django-admin-sortable==2.2.3
django-storages==1.10.1
dnspython==2.0.0
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.10
Pillow==8.0.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.0
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.2.4
urllib3==1.25.11
